Movimentação de logica para a viewModel
