+++ YDMOfflineAccount/Home/ViewController/Extensions/CollectionView/HomeViewController+CollectionView.swift
@@ -11,7 +11,7 @@ extension HomeViewController: UICollectionViewDelegate, UICollectionViewDataSour
   
   func collectionView(_ collectionView: UICollectionView,
                       numberOfItemsInSection section: Int) -> Int {
-    return listItensOffiline.count
+    return viewModel?.listItensOffiline.count ?? 0
   }
   
   func collectionView(
@@ -24,7 +24,9 @@ extension HomeViewController: UICollectionViewDelegate, UICollectionViewDataSour
       for: indexPath
     ) as! HomeViewCell
     
-    cell.populateView(item: (listItensOffiline[indexPath.row]))
+    if let item = viewModel?.listItensOffiline[indexPath.row] {
+      cell.populateView(item: item)
+    }
     
     return cell
   }
@@ -42,7 +44,9 @@ extension HomeViewController: UICollectionViewDelegate, UICollectionViewDataSour
       for: indexPath
     ) as! HomeHeaderViewCell
     
-    header.populateView(user: viewModel!.currentUser)
+    if let user = viewModel?.currentUser {
+      header.populateView(user: user)
+    }
     
     return header
     
@@ -64,7 +68,9 @@ extension HomeViewController: UICollectionViewDelegateFlowLayout {
   }
   
   func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
-    viewModel?.onCard(tag: listItensOffiline[indexPath.row].type)
+    if let item = viewModel?.listItensOffiline[indexPath.row].type {
+      viewModel?.onCard(tag: item)
+    }
   }
 
 }
